refactor group details edit & add toast
